SQL-605: Move tests into new file
