+++ odbc/src/lib.rs
@@ -2,9 +2,4 @@ mod api;
 pub use api::*;
 
 #[cfg(test)]
-mod tests {
-    #[test]
-    fn it_works() {
-        // no-op
-    }
-}
+mod tests;
